globbing file clean up
